fix: open handlers for jest
