+++ package.json
@@ -27,7 +27,7 @@
     "prestart": "npm run build",
     "start:dev": "NODE_ENV=development nodemon",
     "start": "NODE_ENV=production node dist/index.js",
-    "test": "NODE_ENV=test jest",
+    "test": "NODE_ENV=test jest --runInBand --detectOpenHandles",
     "coverage": "NODE_ENV=test jest --coverage"
   }
 }
