+++ src/schema.ts
@@ -1,0 +1,23 @@
+const sqlite3 = require('sqlite3').verbose()
+export const db = new sqlite3.Database(':memory:')
+
+export const buildSchemas = (dbInsance: any) => {
+  const createRideTableSchema = `
+    CREATE TABLE Rides
+    (
+    rideID INTEGER PRIMARY KEY AUTOINCREMENT,
+    startLat DECIMAL NOT NULL,
+    startLong DECIMAL NOT NULL,
+    endLat DECIMAL NOT NULL,
+    endLong DECIMAL NOT NULL,
+    riderName TEXT NOT NULL,
+    driverName TEXT NOT NULL,
+    driverVehicle TEXT NOT NULL,
+    created DATETIME default CURRENT_TIMESTAMP
+    )
+  `
+  dbInsance.run(createRideTableSchema)
+  return db
+}
+
+
